Opus/qa-agent/requirements.txt
requests
pyyaml
supabase
anthropic
colorama
